requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
cytoolz==0.9.0.1
eth-abi==0.5.0
eth-keys==0.1.0b4
eth-tester==0.1.0b11
eth-utils==0.8.1
Flask==0.12.2
Flask-Script==2.0.6
idna==2.6
itsdangerous==0.24
Jinja2==2.10
lru-dict==1.1.6
MarkupSafe==1.0
pika==0.11.2
pysha3==1.0.2
requests==2.18.4
rlp==0.6.0
semantic-version==2.6.0
six==1.11.0
toolz==0.9.0
urllib3==1.22
web3==3.16.5
websocket-client==0.47.0
Werkzeug==0.14.1
